# docs/swagger.yaml
definitions:
  handler.SoftUpdateRequestDTO:
    properties:
      idade:
        type: integer
      sobrenome:
        type: string
    required:
    - idade
    - sobrenome
    type: object
  handler.request:
    properties:
      altura:
        type: integer
      ativo:
        type: boolean
      dataDeCriacao:
        type: string
      email:
        type: string
      idade:
        type: integer
      nome:
        type: string
      sobrenome:
        type: string
    required:
    - altura
    - ativo
    - dataDeCriacao
    - email
    - idade
    - nome
    - sobrenome
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle Go Web exercises.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Go Web Meli Exercises
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: get users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List users
      tags:
      - users
  /users/:id:
    delete:
      consumes:
      - application/json
      description: delete users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete users
      tags:
      - Users
  /users/create:
    post:
      consumes:
      - application/json
      description: Create users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: user to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Create users
      tags:
      - Users
  /users/softupdate/:id:
    patch:
      consumes:
      - application/json
      description: updates users sobrenome and idade fields
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: sobrenome to update
        in: body
        name: softUpdateDTO
        required: true
        schema:
          $ref: '#/definitions/handler.SoftUpdateRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: SoftUpdate users
      tags:
      - Users
  /users/update/:id:
    put:
      consumes:
      - application/json
      description: update users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update users
      tags:
      - Users
swagger: "2.0"